refactor toward add statements
