+++ policyengine_us/tests/policy/baseline/gov/states/ct/tax/income/ct_income_tax_main_rates.yaml
@@ -4,10 +4,8 @@
     state_code: CT
     ct_taxable_income: 1_000
     filing_status: SINGLE
-    ct_income_tax_phase_out_add_back: 0
-    ct_income_tax_recapture: 0
   output:
-    ct_income_tax_before_personal_credit: 30
+    ct_income_tax_main_rates: 30
 
 - name: 2022 Separate household with $20,000 income with some add back, no credits
   period: 2022
@@ -15,10 +13,8 @@
     state_code: CT
     ct_taxable_income: 20_000
     filing_status: SEPARATE
-    ct_income_tax_phase_out_add_back: 100
-    ct_income_tax_recapture: 200
   output:
-    ct_income_tax_before_personal_credit: 1_100
+    ct_income_tax_main_rates: 1_100
 
 - name: 2024 Single household with $1,000 income and no add back or credits
   period: 2024
@@ -26,10 +22,8 @@
     state_code: CT
     ct_taxable_income: 1_000
     filing_status: SINGLE
-    ct_income_tax_phase_out_add_back: 0
-    ct_income_tax_recapture: 0
   output:
-    ct_income_tax_before_personal_credit: 20
+    ct_income_tax_main_rates: 20
 
 - name: 2024 Separate household with $20,000 income with some add back, no credits
   period: 2024
@@ -37,7 +31,5 @@
     state_code: CT
     ct_taxable_income: 20_000
     filing_status: SEPARATE
-    ct_income_tax_phase_out_add_back: 100
-    ct_income_tax_recapture: 200
   output:
-    ct_income_tax_before_personal_credit: 950
+    ct_income_tax_main_rates: 950
